refactor: members/upload 폴더 구조 개선
