+++ components/members/upload/sections/TmiSection/index.tsx
@@ -5,8 +5,8 @@ import TextArea from '@/components/common/TextArea';
 import MemberFormHeader from '@/components/members/upload/forms/FormHeader';
 import MemberFormItem from '@/components/members/upload/forms/FormItem';
 import { MemberFormSection } from '@/components/members/upload/forms/FormSection';
-import FavorToggle from '@/components/members/upload/TmiSection/FavorToggle';
-import MbtiSelector from '@/components/members/upload/TmiSection/MbtiSelector';
+import FavorToggle from '@/components/members/upload/sections/TmiSection/FavorToggle';
+import MbtiSelector from '@/components/members/upload/sections/TmiSection/MbtiSelector';
 import {
   FavorAlcohol,
   FavorFishBread,
@@ -15,7 +15,7 @@ import {
   FavorSweetAndSourPork,
   FavorTteokbokki,
   Mbti,
-} from '@/components/members/upload/TmiSection/types';
+} from '@/components/members/upload/sections/TmiSection/types';
 import { MemberUploadForm } from '@/components/members/upload/types';
 
 export default function TmiSection() {
